+++ PrototypeSneaking/Assets/PrototypeSneaking/Resources/Prefabs/Enemy.prefab
@@ -571,6 +571,7 @@ GameObject:
   - component: {fileID: 1303873116699523012}
   - component: {fileID: 2483953759257903524}
   - component: {fileID: 4833989362533742663}
+  - component: {fileID: 915730001}
   m_Layer: 0
   m_Name: Enemy
   m_TagString: Untagged
@@ -624,7 +625,7 @@ Rigidbody:
   m_UseGravity: 1
   m_IsKinematic: 0
   m_Interpolate: 0
-  m_Constraints: 4
+  m_Constraints: 36
   m_CollisionDetection: 0
 --- !u!114 &1303873116699523012
 MonoBehaviour:
@@ -652,7 +653,7 @@ NavMeshAgent:
   m_Enabled: 1
   m_AgentTypeID: 0
   m_Radius: 0.4
-  m_Speed: 3.5
+  m_Speed: 2
   m_Acceleration: 8
   avoidancePriority: 50
   m_AngularSpeed: 0
@@ -677,7 +678,24 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   attackDistance: 1
-  rotationSpeed: 0.2
+  rotationSpeed: 0.5
+  wonderHabit: {fileID: 915730001}
+--- !u!114 &915730001
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7402465664982925360}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4d3dfab5b4c094af2b7c0f8f1085d078, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  wonderingPaths:
+  - {x: 4, y: 0, z: 2}
+  - {x: 1, y: 0, z: 2}
+  rotationSpeed: 0.5
 --- !u!1 &8056460807633106134
 GameObject:
   m_ObjectHideFlags: 0
