--- PrototypeSneaking/Assets/PrototypeSneaking/Scripts/Domain/Stage/CharacterHabit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f7e96a79e63624ec5a60f063b4a6bc06
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
