+++ PrototypeSneaking/Assets/PrototypeSneaking/Scripts/Domain/Stage/Habit/CharacterHabit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 649a7c168450049989330c18c44b2f07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
