+++ PrototypeSneaking/Assets/PrototypeSneaking/Scripts/Domain/Stage/Habit/LinearTrackingHabit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb906a32a1a9d44799f263b38e029c00
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
