+++ PrototypeSneaking/Assets/PrototypeSneaking/Scripts/Domain/Stage/Habit/WonderHabit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d3dfab5b4c094af2b7c0f8f1085d078
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
