--- PrototypeSneaking/Assets/PrototypeSneaking/Scripts/Domain/Stage/TrackingHabit.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 31ba2730ef1c24c1e81911d2bf2df9c6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
