--- PrototypeSneaking/Assets/PrototypeSneaking/Scripts/Domain/Stage/TrackingHabit/TrackingHabit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5d12c6b3446d446d1b94efb87781593f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
