+++ README.md
@@ -1,3 +1,18 @@
-# Prototype Sneaking
+# プロトタイプステルスゲーム
+
+1ヶ月弱で作ろうシリーズ
+
+毎日ちょこちょこやって1ヶ月で作れればいいなーって感じ
+
+# 成果物
+
+相手の視界に入ったら追跡される
+
+相手の視線を避けてゴール（画面右）を目指す
+
+![](./doc/img/tracking.gif)
+
+# 設計
 
 [設計](./PROGRAM_DESIGN.md)
+
